Stop clobbering modifiers for digit keypresses
Previously, when `0` was used to call `line_home`, this would clobber
the modifiers even if a digit 1-9 was pressed before `0`, which should
signify the beginning of a modifier key sequence and cause `0` to behave
as a modifier until the next command is pressed.

This checks if a current modifier is in effect, and if so it does not
treat any digit keypresses as commands until the modifier is cleared.