Support reading from stdin (fixes #79)
In bin/tabview, support reading directly from stdin if the filename is '-'. In
that case, buffer the entire data into a BytesIO stream, then fix stdin by
re-opening the controlling terminal directly (assigning to sys.stdin is not
enough for the curses module).

In the tabview.view, support reading also from an IOBase object directly.
Rework the reload loop in order to ignore (for now) open/reload errors instead
of failing.